+++ src/main/platforms-config.js
@@ -91,7 +91,8 @@ const HTML_FILTERING_MODIFIER_PATTERNS = [
  *
  * Pattern parts:
  * - `\\$` — modifiers divider
- * - `(?!((path|domain)=.*])).*` — negative lookahead to exclude rules like `[$path=...]##.textad,protobuf`
+ * - `(?!#|(path|domain)=.*]|.*removeparam=).*` — negative lookahead to exclude CSS rules (#$#) and rules like
+ *   `[$path=...]##.textad,protobuf` and rules with `$removeparam` modifier like `$removeparam=protobuf`
  * - `.*protobuf` — protobuf modifier itself
  * - `(,|=|$)` — end of line or modifiers divider, as `$protobuf` can be followed by other modifiers (`,`),
  *   it may have a value (`=`), or it may be the last modifier in the rule (`$`).
@@ -100,7 +101,7 @@ const HTML_FILTERING_MODIFIER_PATTERNS = [
  * https://github.com/AdguardTeam/CoreLibs/issues/1778
  */
 const PROTOBUF_MODIFIER_PATTERNS = [
-    '\\$(?!((path|domain)=.*])).*protobuf(,|=|$)',
+    '\\$(?!#|(path|domain)=.*]|.*removeparam=).*protobuf(,|=|$)',
 ];
 
 /**
@@ -109,14 +110,15 @@ const PROTOBUF_MODIFIER_PATTERNS = [
  *
  * Pattern parts:
  * - `\\$` — modifiers divider
- * - `(?!((path|domain)=.*])).*` — negative lookahead to exclude rules like `[$path=...]##.textad,[app="ads"]`
+ * - `(?!#|(path|domain)=.*]|.*removeparam=).*` — negative lookahead to exclude CSS rules (#$#) and rules like
+ *   `[$path=...]##.textad,[app="ads"]` and rules with `$removeparam` modifier like `$removeparam=app=ads`
  * - `.*app=` — app= modifier itself
  *
  * @example
  * ```@@||imasdk.googleapis.com^$app=tv.htv.app```
  */
 const APP_MODIFIER_PATTERNS = [
-    '\\$(?!((path|domain)=.*])).*app=',
+    '\\$(?!#|(path|domain)=.*]|.*removeparam=).*app=',
 ];
 
 /**
@@ -125,7 +127,8 @@ const APP_MODIFIER_PATTERNS = [
  *
  * Pattern parts:
  * - `\\$` — modifiers divider
- * - `(?!((path|domain)=.*])).*` — negative lookahead to exclude rules like `[$path=...]##.textad,extension`
+ * - `(?!#|(path|domain)=.*]|.*removeparam=).*` — negative lookahead to exclude CSS rules (#$#) and rules like
+ *   `[$path=...]##.textad,extension` and rules with `$removeparam` modifier like `$removeparam=extension`
  * - `.*extension` — extension modifier itself
  * - `(,|=|$)` — end of line or modifiers divider, as `$extension` can be followed by other modifiers (`,`),
  *   it may have a value (`=`), or it may be the last modifier in the rule (`$`).
@@ -134,7 +137,7 @@ const APP_MODIFIER_PATTERNS = [
  * ```@@||radar.cloudflare.com^$elemhide,extension,content```
  */
 const EXTENSION_MODIFIER_PATTERNS = [
-    '\\$(?!((path|domain)=.*])).*extension(,|=|$)',
+    '\\$(?!#|(path|domain)=.*]|.*removeparam=).*extension(,|=|$)',
 ];
 
 /**
@@ -153,7 +156,8 @@ const ONLY_CONTENT_MODIFIER_PATTERNS = [
  *
  * Pattern parts:
  * - `\\$` — modifiers divider
- * - `(?!((path|domain)=.*])).*` — negative lookahead to exclude rules like `[$path=...]##.textad,content`
+ * - `(?!#|(path|domain)=.*]|.*removeparam=).*` — negative lookahead to exclude CSS rules (#$#) and rules like
+ *   `[$path=...]##.textad,content` and rules with `$removeparam` modifier like `$removeparam=content`
  * - `.*content` — content modifier itself
  * - `(,|$)` — end of line or modifiers divider, as `$content` can be followed by other modifiers (`,`).
  *
@@ -161,7 +165,7 @@ const ONLY_CONTENT_MODIFIER_PATTERNS = [
  * ```@@||dnsleaktest.com^$content,elemhide,jsinject```
  */
 const CONTENT_MODIFIER_PATTERNS = [
-    '\\$(?!((path|domain)=.*])).*content(,|$)',
+    '\\$(?!#|(path|domain)=.*]|.*removeparam=).*content(,|$)',
 ];
 
 /**
@@ -170,7 +174,8 @@ const CONTENT_MODIFIER_PATTERNS = [
  *
  * Pattern parts:
  * - `\\$` — modifiers divider
- * - `(?!((path|domain)=.*])).*` — negative lookahead to exclude rules like `[$path=...]##.textad,jsinject`
+ * - `(?!#|(path|domain)=.*]|.*removeparam=).*` — negative lookahead to exclude CSS rules (#$#) and rules like
+ *   `[$path=...]##.textad,jsinject` and rules with `$removeparam` modifier like `$removeparam=jsinject`
  * - `.*jsinject` — jsinject modifier itself
  * - `(,|$)` — end of line or modifiers divider, as `$jsinject` can be followed by other modifiers (`,`).
  *
@@ -178,7 +183,7 @@ const CONTENT_MODIFIER_PATTERNS = [
  * ```@@://www.atlassian.com^$elemhide,jsinject,extension```
  */
 const JSINJECT_MODIFIER_PATTERNS = [
-    '\\$(?!((path|domain)=.*])).*jsinject(,|$)',
+    '\\$(?!#|(path|domain)=.*]|.*removeparam=).*jsinject(,|$)',
 ];
 
 /**
@@ -187,7 +192,8 @@ const JSINJECT_MODIFIER_PATTERNS = [
  *
  * Pattern parts:
  * - `\\$` — modifiers divider
- * - `(?!((path|domain)=.*])).*` — negative lookahead to exclude rules like `[$path=...]##.textad,urlblock`
+ * - `(?!#|(path|domain)=.*]|.*removeparam=).*` — negative lookahead to exclude CSS rules (#$#) and rules like
+ *   `[$path=...]##.textad,urlblock` and rules with `$removeparam` modifier like `$removeparam=urlblock`
  * - `.*urlblock` — urlblock modifier itself
  * - `(,|$)` — end of line or modifiers divider, as `$urlblock` can be followed by other modifiers (`,`).
  *
@@ -195,7 +201,7 @@ const JSINJECT_MODIFIER_PATTERNS = [
  * ```@@||google.com/settings/ads/onweb$urlblock```
  */
 const URLBLOCK_MODIFIER_PATTERNS = [
-    '\\$(?!((path|domain)=.*])).*urlblock(,|$)',
+    '\\$(?!#|(path|domain)=.*]|.*removeparam=).*urlblock(,|$)',
 ];
 
 /**
@@ -204,7 +210,8 @@ const URLBLOCK_MODIFIER_PATTERNS = [
  *
  * Pattern parts:
  * - `\\$` — modifiers divider
- * - `(?!((path|domain)=.*])).*` — negative lookahead to exclude rules like `[$path=...]##.textad,referrerpolicy`
+ * - `(?!#|(path|domain)=.*]|.*removeparam=).*` — negative lookahead to exclude CSS rules (#$#) and rules like
+ *   `[$path=...]##.textad,referrerpolicy` and rules with `$removeparam` modifier like `$removeparam=referrerpolicy`
  * - `.*referrerpolicy` — referrerpolicy modifier itself
  * - `(,|=|$)` — end of line or modifiers divider, as `$referrerpolicy` can be followed by other modifiers (`,`),
  *   it may have a value (`=`), or it may be the last modifier in the rule (`$`).
@@ -213,7 +220,7 @@ const URLBLOCK_MODIFIER_PATTERNS = [
  * ```||yallo.tv^$referrerpolicy=origin```
  */
 const REFERRERPOLICY_MODIFIER_PATTERNS = [
-    '\\$(?!((path|domain)=.*])).*referrerpolicy(,|=|$)',
+    '\\$(?!#|(path|domain)=.*]|.*removeparam=).*referrerpolicy(,|=|$)',
 ];
 
 /**
@@ -222,7 +229,8 @@ const REFERRERPOLICY_MODIFIER_PATTERNS = [
  *
  * Pattern parts:
  * - `\\$` — modifiers divider
- * - `(?!((path|domain)=.*])).*` — negative lookahead to exclude rules like `[$path=...]##.textad,[replace="ads"]`
+ * - `(?!#|(path|domain)=.*]|.*removeparam=).*` — negative lookahead to exclude CSS rules (#$#) and rules like
+ *   `[$path=...]##.textad,[replace="ads"]` and rules with `$removeparam` modifier like `$removeparam=replace=ads`
  * - `.*replace` — replace modifier itself
  * - `(,|=|$)` — end of line or modifiers divider, as `$replace` can be followed by other modifiers (`,`),
  *   it may have a value (`=`), or it may be the last modifier in the rule (`$`).
@@ -231,7 +239,7 @@ const REFERRERPOLICY_MODIFIER_PATTERNS = [
  * ```||pubads.g.doubleclick.net/gampad/live/ads?correlator=$replace=/(<VAST[\s\S]*?>)[\s\S]*<\/VAST>/\$1<\/VAST>/```
  */
 const REPLACE_MODIFIER_PATTERNS = [
-    '\\$(?!((path|domain)=.*])).*replace(,|=|$)',
+    '\\$(?!#|(path|domain)=.*]|.*removeparam=).*replace(,|=|$)',
 ];
 
 /* eslint-disable max-len */
@@ -242,7 +250,8 @@ const REPLACE_MODIFIER_PATTERNS = [
  *
  * Pattern parts:
  * - `\\$` — modifiers divider
- * - `(?!((path|domain)=.*])).*` — negative lookahead to exclude rules like `[$path=...]##.textad,[hls="ads"]`
+ * - `(?!#|(path|domain)=.*]|.*removeparam=).*` — negative lookahead to exclude CSS rules (#$#) and rules like
+ *   `[$path=...]##.textad,[hls="ads"]` and rules with `$removeparam` modifier like `$removeparam=hls=ads`
  * - `.*hls` — hls modifier itself
  * - `(,|=|$)` — end of line or modifiers divider, as `$hls` can be followed by other modifiers (`,`),
  *   it may have a value (`=`), or it may be the last modifier in the rule (`$`).
@@ -251,7 +260,7 @@ const REPLACE_MODIFIER_PATTERNS = [
  * ```||pubads.g.doubleclick.net/ondemand/hls/*.m3u8$hls=/redirector\.googlevideo\.com\/videoplayback[\s\S]*?dclk_video_ads/,domain=10play.com.au```
  */
 const HLS_MODIFIER_PATTERNS = [
-    '\\$(?!((path|domain)=.*])).*hls(,|=|$)',
+    '\\$(?!#|(path|domain)=.*]|.*removeparam=).*hls(,|=|$)',
 ];
 
 /**
@@ -260,7 +269,8 @@ const HLS_MODIFIER_PATTERNS = [
  *
  * Pattern parts:
  * - `\\$` — modifiers divider
- * - `(?!((path|domain)=.*])).*` — negative lookahead to exclude rules like `[$path=...]##.textad,[jsonprune="ads"]`
+ * - `(?!#|(path|domain)=.*]|.*removeparam=).*` — negative lookahead to exclude CSS rules (#$#) and rules like
+ *   `[$path=...]##.textad,[jsonprune="ads"]` and rules with `$removeparam` modifier like `$removeparam=jsonprune`
  * - `.*jsonprune` — jsonprune modifier itself
  * - `(,|=|$)` — end of line or modifiers divider, as `$jsonprune` can be followed by other modifiers (`,`),
  *   it may have a value (`=`), or it may be the last modifier in the rule (`$`).
@@ -269,7 +279,7 @@ const HLS_MODIFIER_PATTERNS = [
  * ```.com/watch?v=$xmlhttprequest,jsonprune=\$..[adPlacements\, adSlots\, playerAds],domain=youtubekids.com|youtube-nocookie.com|youtube.com```
  */
 const JSONPRUNE_MODIFIER_PATTERNS = [
-    '\\$(?!((path|domain)=.*])).*jsonprune(,|=|$)',
+    '\\$(?!#|(path|domain)=.*]|.*removeparam=).*jsonprune(,|=|$)',
 ];
 
 /* eslint-enable max-len */
@@ -280,7 +290,8 @@ const JSONPRUNE_MODIFIER_PATTERNS = [
  *
  * Pattern parts:
  * - `\\$` — modifiers divider
- * - `(?!((path|domain)=.*])).*` — negative lookahead to exclude rules like `[$path=...]##.textad,[removeparam="ads"]`
+ * - `(?!#|(path|domain)=.*]).*` — negative lookahead to exclude CSS rules (#$#) and rules like
+ *   `[$path=...]##.textad,[removeparam="ads"]`
  * - `.*removeparam` — removeparam modifier itself
  * - `(,|=|$)` — end of line or modifiers divider, as `$removeparam` can be followed by other modifiers (`,`),
  *   it may have a value (`=`), or it may be the last modifier in the rule (`$`).
@@ -289,7 +300,7 @@ const JSONPRUNE_MODIFIER_PATTERNS = [
  * ```$removeparam=fb_ref```
  */
 const REMOVEPARAM_MODIFIER_PATTERNS = [
-    '\\$(?!((path|domain)=.*])).*removeparam(,|=|$)',
+    '\\$(?!#|(path|domain)=.*]).*removeparam(,|=|$)',
 ];
 
 /**
@@ -298,7 +309,8 @@ const REMOVEPARAM_MODIFIER_PATTERNS = [
  *
  * Pattern parts:
  * - `\\$` — modifiers divider
- * - `(?!((path|domain)=.*])).*` — negative lookahead to exclude rules like `[$path=...]##.textad,[removeheader="ads"]`
+ * - `(?!#|(path|domain)=.*]|.*removeparam=).*` — negative lookahead to exclude CSS rules (#$#) and rules like
+ *   `[$path=...]##.textad,[removeheader="ads"]` and rules with `$removeparam` modifier like `$removeparam=removeheader`
  * - `.*removeheader` — removeheader modifier itself
  * - `(,|=|$)` — end of line or modifiers divider, as `$removeheader` can be followed by other modifiers (`,`),
  *   it may have a value (`=`), or it may be the last modifier in the rule (`$`).
@@ -307,7 +319,7 @@ const REMOVEPARAM_MODIFIER_PATTERNS = [
  * ```||dubznetwork.com^$removeheader=refresh```
  */
 const REMOVEHEADER_MODIFIER_PATTERNS = [
-    '\\$(?!((path|domain)=.*])).*removeheader(,|=|$)',
+    '\\$(?!#|(path|domain)=.*]|.*removeparam=).*removeheader(,|=|$)',
 ];
 
 /**
@@ -316,7 +328,8 @@ const REMOVEHEADER_MODIFIER_PATTERNS = [
  *
  * Pattern parts:
  * - `\\$` — modifiers divider
- * - `(?!((path|domain)=.*])).*` — negative lookahead to exclude rules like `[$path=...]##.textad,[stealth="ads"]`
+ * - `(?!#|(path|domain)=.*]|.*removeparam=).*` — negative lookahead to exclude CSS rules (#$#) and rules like
+ *  `[$path=...]##.textad,[stealth="ads"]` and rules with `$removeparam` modifier like `$removeparam=stealth`
  * - `.*stealth` — stealth modifier itself
  * - `(,|=|$)` — end of line or modifiers divider, as `$stealth` can be followed by other modifiers (`,`),
  *   it may have a value (`=`), or it may be the last modifier in the rule (`$`).
@@ -325,7 +338,7 @@ const REMOVEHEADER_MODIFIER_PATTERNS = [
  * ```@@.php?play_vid=$subdocument,stealth=referrer,domain=xyflv.cc```
  */
 const STEALTH_MODIFIER_PATTERNS = [
-    '\\$(?!((path|domain)=.*])).*stealth(,|=|$)',
+    '\\$(?!#|(path|domain)=.*]|.*removeparam=).*stealth(,|=|$)',
 ];
 
 /**
@@ -334,7 +347,8 @@ const STEALTH_MODIFIER_PATTERNS = [
  *
  * Pattern parts:
  * - `\\$` — modifiers divider
- * - `(?!((path|domain)=.*])).*` — negative lookahead to exclude rules like `[$path=...]##.textad,cookie`
+ * - `(?!#|(path|domain)=.*]|.*removeparam=).*` — negative lookahead to exclude CSS rules (#$#) and rules like
+ *   `[$path=...]##.textad,cookie` and rules with `$removeparam` modifier like `$removeparam=cookie`
  * - `.*cookie` — cookie modifier itself
  * - `(,|=|$)` — end of line or modifiers divider, as `$cookie` can be followed by other modifiers (`,`),
  *   it may have a value (`=`), or it may be the last modifier in the rule (`$`).
@@ -343,7 +357,7 @@ const STEALTH_MODIFIER_PATTERNS = [
  * ```$cookie=_ga```
  */
 const COOKIE_MODIFIER_PATTERNS = [
-    '\\$(?!((path|domain)=.*])).*cookie(,|=|$)',
+    '\\$(?!#|(path|domain)=.*]|.*removeparam=).*cookie(,|=|$)',
 ];
 
 /**
@@ -352,7 +366,8 @@ const COOKIE_MODIFIER_PATTERNS = [
  *
  * Pattern parts:
  * - `\\$` — modifiers divider
- * - `(?!((path|domain)=.*])).*` — negative lookahead to exclude rules like `[$path=...]##.textad,redirect`
+ * - `(?!#|(path|domain)=.*]|.*removeparam=).*` — negative lookahead to exclude CSS rules (#$#) and rules like
+ *   `[$path=...]##.textad,redirect` and rules with `$removeparam` modifier like `$removeparam=redirect`
  * - `.*redirect` — redirect modifier itself
  * - `(,|=|$)` — end of line or modifiers divider, as `$redirect` can be followed by other modifiers (`,`),
  *   it may have a value (`=`), or it may be the last modifier in the rule (`$`).
@@ -361,7 +376,7 @@ const COOKIE_MODIFIER_PATTERNS = [
  * ```||google-analytics.com/analytics.js$script,redirect=google-analytics,domain=~olx.*|~banki.ru|~bigc.co.th```
  */
 const REDIRECT_MODIFIER_PATTERNS = [
-    '\\$(?!((path|domain)=.*])).*redirect(,|=|$)',
+    '\\$(?!#|(path|domain)=.*]|.*removeparam=).*redirect(,|=|$)',
 ];
 
 /**
@@ -370,7 +385,8 @@ const REDIRECT_MODIFIER_PATTERNS = [
  *
  * Pattern parts:
  * - `\\$` — modifiers divider
- * - `(?!((path|domain)=.*])).*` — negative lookahead to exclude rules like `[$path=...]##.textad,redirect-rule`
+ * - `(?!#|(path|domain)=.*]|.*removeparam=).*` — negative lookahead to exclude CSS rules (#$#) and rules like
+ *   `[$path=...]##.textad,redirect-rule` and rules with `$removeparam` modifier like `$removeparam=redirect-rule`
  * - `.*redirect-rule` — redirect-rule modifier itself
  * - `(,|=|$)` — end of line or modifiers divider, as `$redirect-rule` can be followed by other modifiers (`,`),
  *   it may have a value (`=`), or it may be the last modifier in the rule (`$`).
@@ -379,7 +395,7 @@ const REDIRECT_MODIFIER_PATTERNS = [
  * ```$script,third-party,redirect-rule=noopjs,domain=paraphraser.io```
  */
 const REDIRECT_RULE_MODIFIER_PATTERNS = [
-    '\\$(?!((path|domain)=.*])).*redirect-rule(,|=|$)',
+    '\\$(?!#|(path|domain)=.*]|.*removeparam=).*redirect-rule(,|=|$)',
 ];
 
 /**
@@ -388,7 +404,8 @@ const REDIRECT_RULE_MODIFIER_PATTERNS = [
  *
  * Pattern parts:
  * - `\\$` — modifiers divider
- * - `(?!((path|domain)=.*])).*` — negative lookahead to exclude rules like `[$path=...]##.textad,empty`
+ * - `(?!#|(path|domain)=.*]|.*removeparam=).*` — negative lookahead to exclude CSS rules (#$#) and rules like
+ *   `[$path=...]##.textad,empty` and rules with `$removeparam` modifier like `$removeparam=empty`
  * - `.*empty` — empty modifier itself
  * - `(,|$)` — end of line or modifiers divider, as `$empty` can be followed by other modifiers (`,`).
  *
@@ -396,7 +413,7 @@ const REDIRECT_RULE_MODIFIER_PATTERNS = [
  * ```Deprecated, use $redirect=nooptext instead```
  */
 const EMPTY_MODIFIER_PATTERNS = [
-    '\\$(?!((path|domain)=.*])).*empty(,|$)',
+    '\\$(?!#|(path|domain)=.*]|.*removeparam=).*empty(,|$)',
 ];
 
 /**